+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### v4.10.x
   - [EE] Fix `options.isDebugActive = true`: do not return internal (not visible by user) tags
+  - [EE] Fix crash when the socket is broken while the stream is still in progress for `GET /template/:templateId` and `GET /render/:reportId`
+
 ### v4.10.2
   - [EE] Fixed `HEAD /render/:renderId`: The request don't delete the generated document anymore.
   - [EE] Fixed `GET /template/:templateId`: If the template doesn't exist, the statusCode `404` is returned instead of `400`.